+++ feature-move-io/src/main/java/me/saiintbrisson/minecraft/MoveInOutFeature.java
@@ -5,6 +5,7 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 
+@SuppressWarnings("unused")
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class MoveInOutFeature implements Feature<Void, Void> {